move files to process and write tests
